Move getDefaultSclangExecutable to util
