+++ util.js
@@ -1,4 +1,23 @@
+const fs = require('fs');
+const path = require('path');
 const vscode = require('vscode');
+
+function getDefaultSclangExecutable() {
+  switch (process.platform) {
+    case 'darwin':
+      return '/Applications/SuperCollider.app/Contents/MacOS/sclang';
+    case 'win32':
+      const root = 'C:\\Program Files';
+      const directories = fs.readdirSync(root, { withFileTypes: true })
+        .filter((entry) => entry.isDirectory() && entry.name.startsWith("SuperCollider-"));
+      if (directories.length > 0) {
+        return path.join(root, directories[0].name, 'sclang.exe');
+      }
+      return 'sclang';
+    default:
+      return 'sclang';
+  }
+}
 
 function stringifyError(value) {
   const { type, error } = value;
@@ -35,4 +54,5 @@ function flashHighlight(editor, range, duration = 250) {
 module.exports = {
   flashHighlight,
   stringifyError,
+  getDefaultSclangExecutable,
 };
